Refactor factor_interaction compat
